BIN2YAP: now supports encrypted yap file output. bootloader: now supports encrypted yap files (-O1 required to reduce binary down to less than 32k). IGPS:definition changed for visualstudio . parse_log: minor changes.

+++ HAL/Interface/IGPS.h
@@ -15,8 +15,8 @@ namespace devices
 		uint8_t satelite_in_view;
 		uint8_t satelite_in_use;
 		uint8_t declination; 			// Magnetic variation in 0.01 degrees (Easterly var. subtracts from true course)
-		unsigned sig				: 4;// GPS quality indicator (0 = Invalid; 1 = Fix; 2 = Differential, 3 = Sensitive)
-		unsigned fix				: 4;// Operating mode, used for navigation (1 = Fix not available; 2 = 2D; 3 = 3D)
+		uint8_t sig				: 4;// GPS quality indicator (0 = Invalid; 1 = Fix; 2 = Differential, 3 = Sensitive)
+		uint8_t fix				: 4;// Operating mode, used for navigation (1 = Fix not available; 2 = 2D; 3 = 3D)
 		float position_accuracy_horizontal;	// unit: variance in meter
 		float position_accuracy_vertical;
 		float velocity_accuracy_horizontal;	// unit: variance in meter/s
